feat: added functionality to the form
