+++ src/assets/styles/globals.css
@@ -107,3 +107,6 @@ body .css-tj5bde-Svg {
   width: 24px;
   height: 24px;
 }
+body .css-100ebu3-menu {
+  top: 56px;
+}
