+++ src/components/Button/Button.tsx
@@ -1,9 +1,9 @@
 'use client';
 import React, { useState } from 'react';
 import { twMerge } from 'tailwind-merge';
-import { Modal } from '..';
 import useVisibility from '@/hooks/useVisibility';
-import Form_Thumb from '@/modules/Form/Form';
+import { Modal } from '..';
+import { Form } from '@/modules';
 
 const Button: React.FC = () => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
@@ -24,7 +24,7 @@ const Button: React.FC = () => {
       </button>
       {isOpen && (
         <Modal isOpen={isOpen} onClose={toggleModal}>
-          <Form_Thumb />
+          <Form />
         </Modal>
       )}
     </>
