+++ src/modules/Form/components/index.tsx
@@ -1,6 +1,5 @@
-import { Submit_form } from './Submit_form';
-import { Accept_form } from './Accept_form';
-import { SelectWrapper } from './Select';
-import { Input } from './Input';
+import Success_modal from './Success_modal';
+import SelectWrapper from './Select';
+import Input from './Input';
 
-export { Submit_form, SelectWrapper,Input, Accept_form };
+export { SelectWrapper, Input, Success_modal };
