+++ src/modules/Form/helpers/saveDataToLocal.tsx
@@ -1,4 +1,4 @@
-import { FormProps } from '../components/Submit_form';
+import { FormProps } from '../Form';
 
 export const saveDataToLocal = (values: FormProps): void => {
   localStorage.setItem('formData', JSON.stringify(values));
